requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
appdirs==1.4.3
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via pytest
backports.os==0.1.1
certifi==2019.3.9
chardet==3.0.4
click==7.0                # via code-annotations
code-annotations==0.3
coverage==4.5.3           # via pytest-cov
django-model-utils==3.1.2
django-waffle==0.15.1
djangorestframework-jwt==1.11.0
djangorestframework==3.9.2
edx-django-utils==1.0.3
edx-drf-extensions==2.1.0
edx-opaque-keys==0.4.4
enum34==1.1.6
fs==2.4.4
funcsigs==1.0.2           # via pytest
future==0.17.1
idna==2.8
jinja2==2.10              # via code-annotations
lxml==4.3.2
markupsafe==1.1.1
more-itertools==5.0.0     # via pytest
newrelic==4.14.0.115
pathlib2==2.3.3           # via pytest, pytest-django
pbr==5.1.3
pluggy==0.9.0             # via pytest
psutil==1.2.1
py==1.8.0                 # via pytest
pycryptodomex==3.7.3
pyjwkest==1.3.2
pyjwt==1.7.1
pymongo==3.7.2
pytest-cov==2.6.1
pytest-django==3.4.8
pytest==4.3.1             # via pytest-cov, pytest-django
python-dateutil==2.8.0
python-slugify==3.0.0     # via code-annotations
pytz==2018.9
pyyaml==5.1
requests==2.21.0
rest-condition==1.0.3
scandir==1.10.0           # via pathlib2
semantic-version==2.6.0
six==1.12.0
stevedore==1.30.1
text-unidecode==1.2       # via python-slugify
typing==3.6.6
urllib3==1.24.1
web-fragments==0.3.0
webob==1.8.5
xblock==1.2.2
